initial file created using this prompt:
I have a folder filled with dashcam footage, broken up into 3 minute or so videos, that sometimes are of a longer trip (say, an hour), that are uninterrupted, and some that are shorter. Each video has a filename something like this: 2024_0921_220728_003 2024_0921_220427_002, the 002 or 003 or etc is just "which number video", that's a bit less important. Before that, you likely recognize the date. I currently have the ffmpeg python package installed.

Based on the timestamps in the video titles, I want a python script that will "concatenate related sessions". To do so, start with the earliest video in the directory, based on the filename only. You can assume the target directory contains only videos of interest. Then, figure out how long the video is, and see if the next timestamped file is within 1 minute from the end of the prior file. If so, make sure to put them together, but look forward until you "run out of videos that are aligned in this way". If you find a jump that's more than 5 minutes long, you can assume that's a different driving session, repeat this until we process every file in the folder. However, for this initial draft, read the video to find the video length using the appropriate library, but don't actually create the finished videos just yet. First just make a target output folder path, (create if doesn't exist,), and fill it with text files that contain the filename that's just the "starts" of every session, and inside write out each file you would have put together