doesn't work just yet, but intermediate code was generated with this prompt once I moved along to another conversation
I have some python code here that takes dashcam footage that's in an input folder, and based on the video information concatenates "the same driving session" together, since dashcams by default save in a couple minute clips. Currently, its just in "debug mode" and just prints a text (so I can check its combining what seems sensible before I work with gigs of video data). I have confirmed that the proper files are getting accessed via the output. Keep all that as is, but use the "new file name" for the actual concatenation which I am requesting that you write. Access the requested files from each text file (you don't actually have to read the created text file, use the intermediate variables, I have confirmed its correct), and simply do a "copy paste" via the video editor, don't reprocess it or anything, for speed. If it is possible to do it with ffmpeg and not require an additional package do that. Furthermore, provide the complete code for me and leave nothing out