initial generation with this prompt
I have a directory that only contains video files in mp4 format. They are all dashcam footages, and the filename is the exact start time of the video. 2024_0921_222828_010 is one example of such a filename (likely has a .mp4 at the end if you are directly accessing directory files via code). The last underscore with a number is just the next video recorded in series, let's ignore that number. (_010 in this case). I want a copy of merged videos (with proper metadata reflecting new longer lengths) with "the same trip" being combined together instead of remaining as separate files. I am trying to write a complete python script that will do a "basic copy and paste" from the directory that contains my original files, to another directory. I have provided the paths below, leave them at the beginning of the script for simplicity:

# Hardcoded directories

source_dir = r"D:\Patrick's Documents\Other Things\Video\Dashcam\Hawaii Trip"

output_dir = r"D:\Patrick's Documents\Other Things\Video\Dashcam\concatenated_output" 

Here are the steps that you can use to place files into my output_dir.

First, read through every mp4 file in that folder, and figure out the time of creation of each one. You can assume they will be sequential, if they are not, sort them from oldest to newest. Create a "user defined hardcoded variable" for "maximum length in minutes of each video". Let's prefill it with 4 minutes, (since most videos are 3 minutes long, but let's make it flexible to change at the top in case this changes). Starting from the first video, check when the next video in the sequence was created. If that time is LESS than the video_length_max defined by the user, add it to sequence. Go to that video, and repeat the process on the next one in the sequence. If the next video takes place at a time LONGER than that, it is a new session. For each merged video, title it the same as the first video in the sequence.